Don't throw exception if the header has authorization and a token set to undefined
A common patter with Single Pages Applications is to add an http interceptor that would add the Authorization header to all calls even those that are not meant to be under auth. They just set an 'undefined' token and lets the credentials do the rest.
This change allows such request to pass through the same way as if they don't had the Authorization header.